fix: change the data type of variables

+++ COTP/schastlivye_nomera.cpp
@@ -5,9 +5,9 @@ int main() {
 	ios::sync_with_stdio(false);
 	int n;
 	cin >> n;
-	int sum = 0;
+	long long sum = 0;
 	for (int i = 1; i <= n; i++)
-		sum += 1 << i;
+		sum += 1ll << i;
 	cout << sum;
 	return 0;
 }
